Fix sonar qube complains

+++ src/main/java/org/folio/des/scheduling/acquisition/AcqBaseExportTaskTrigger.java
@@ -145,11 +145,11 @@ public class AcqBaseExportTaskTrigger extends AbstractExportTaskTrigger {
                                      List<DayOfWeek> sortedChosenDays,
                                      DayOfWeek currentDay,
                                      Integer weeksFrequency) {
-    long daysToSunday = DayOfWeek.SUNDAY.getValue() - currentDay.getValue();
+    long daysToSunday = (long) DayOfWeek.SUNDAY.getValue() - currentDay.getValue();
     DayOfWeek firstDayFromNextWeek = sortedChosenDays.get(0);
     return scheduleDateTime
       .plusDays(daysToSunday + firstDayFromNextWeek.getValue())
-      .plusWeeks(weeksFrequency - 1);
+      .plusWeeks(weeksFrequency - 1L);
   }
 
   @SneakyThrows
